gralloc3: [g19] set BLOB format stride same as alloc width
Previously stride value of BLOB format was set to 0,
as stride has no meaning for BLOB formats.

However, this value must be set to be the same as the alloc
width of the buffer (which is also equal to the length of
the buffer).

Otherwise the following CTS test fails:
CtsNNAPITestCases -t MemoryTest#TestAHardwareBuffer

Change-Id: Id7b6267c09bbed398eb6984e3c9071cb455f4a2a
Signed-off-by: SeYeong Byeon <sy.byeon@samsung.com>