gralloc4: Enable SBWC when vpu is specified as consumer
There is still an architecture problem in media frameworks that the
buffer usage is limited to 32 bit. It results the MFC IP flag cannot be
successfully passed to Gralloc module. Before Media frameworks resolves
the limitation in usage flag, the patch considers MFC as the only
consumer when VIDEO_ENCODER is specified.

Bug: b/163821791
Test: Enable Front Cam and Rear Cam
Test: 3P App compatibility test.

Change-Id: I404d49ed33762d3b621523761be2c8bb7c383498