Disable CPU alignment for Camera's RAW10 when not explicitly specified
A RAW10 buffer gets 40 bits/5 bytes CPU aligned. Combining that with
other 64 bytes alignment requirement, makes the alignment 320. This is
unexpected for the camera. Disable CPU alignment when CPU usage is not
specified for now.

A proper alignment fix should align only the base number of width pixels
(not after aligning the pixels themselves) and then during buffer map to
CPU, it should map only the required width rather than the whole size of
dmabuf. But that requires larger change and can have other
ramifications.

Bug: 235052149
Test: Recommended shot works fine
Change-Id: I546ac3bad491e4950d192f8b3f4a20bc3ff5534f